requirements.txt
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
nextcord==2.0.0a10
python-dotenv==0.20.0
requests==2.27.1
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.9
yarl==1.7.2
